+++ test/smtp-connection-test.js
@@ -1384,6 +1384,65 @@ describe('SMTPServer', function() {
         });
     });
 
+    describe('Secure PROXY server', function() {
+        let PORT = 1336;
+
+        let server = new SMTPServer({
+            maxClients: 5,
+            logger: false,
+            useProxy: true,
+            secure: true,
+            onConnect(session, callback) {
+                if (session.remoteAddress === '1.2.3.4') {
+                    let err = new Error('Blacklisted IP');
+                    err.responseCode = 421;
+                    return callback(err);
+                }
+                callback();
+            }
+        });
+
+        beforeEach(function(done) {
+            server.listen(PORT, '127.0.0.1', done);
+        });
+
+        afterEach(function(done) {
+            server.close(done);
+        });
+
+        it('should rewrite remote address value', function(done) {
+            let connection = new Client({
+                port: PORT,
+                host: '127.0.0.1',
+                tls: {
+                    rejectUnauthorized: false
+                }
+            });
+
+            connection.on('end', done);
+
+            connection.connect(function() {
+                let conn;
+                // get first connection
+                server.connections.forEach(function(val) {
+                    if (!conn) {
+                        conn = val;
+                    }
+                });
+                // default remote address should be overriden by the value from the PROXY header
+                expect(conn.remoteAddress).to.equal('198.51.100.22');
+                expect(conn.remotePort).to.equal(35646);
+                connection.quit();
+            });
+
+            connection._socket.write('PROXY TCP4 198.51.100.22 203.0.113.7 35646 80\r\n');
+            connection._upgradeConnection(err => {
+                expect(err).to.not.exist;
+                // server should respond with greeting after this point
+            });
+        });
+    });
+
     describe('onClose handler', function() {
         let PORT = 1336;